Ensure cookie gets deleted when user logs out
For some reason, when testing this with the `mymove` repo, the
cookie wasn't being deleted. Adding these additional cookie settings
allowed the cookie to be deleted.

Looking at how other Golang session management packages delete cookies,
it seems to be a standard thing to have all of these cookie attributes
in place. To delete the cookie, the `Value` is set to an empty
string, `MaxAge` to `-1`, and `Expires` to `time.Unix(1, 0)`.